build(molecule): add all 4 target platforms
The molecule tests now feature all four platforms that we can test using
containers. Testing CentOS-7 in containers is no easy task, therefore
this is dropped. Nevertheless we try to support CentOS-7 as long we have
servers with this OS. Testing is done manually locally using VirtualBox
on a best effort basis.

In addition the latest suggestions from our template are implemented
now, e.g. having a default scenario for GH and a local scenario for
local testing. The playbooks are the same for both scenarios and
symlinked. Only the molecule configuration differs for the time being.